# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GridShield AI API",
    "description": "Explainable Cyber-Physical Resilience Platform for Electrical Power Grids (Research/Educational Digital Twin)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "summary": "Create Run",
        "description": "Executes a complete digital twin run through the full resilience loop:\nSIMULATION -> TELEMETRY -> DETECTION -> ATTRIBUTION -> RISK -> INCIDENT -> MITIGATION",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "description": "Returns details and timeline of a specific run.",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/stream": {
      "get": {
        "summary": "Stream Run",
        "description": "Server-Sent Events (SSE) live telemetry and event stream.",
        "operationId": "stream_run_api_v1_runs__run_id__stream_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "summary": "Get Incident Detail",
        "description": "Returns complete details of a specific incident.",
        "operationId": "get_incident_detail_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/simulate-impact": {
      "post": {
        "summary": "Post Simulate Impact",
        "description": "Runs forward consequence simulation ('if ignored').",
        "operationId": "post_simulate_impact_api_v1_incidents__incident_id__simulate_impact_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          },
          {
            "name": "projected_steps",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 5,
              "default": 30,
              "title": "Projected Steps"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/recommend-mitigation": {
      "post": {
        "summary": "Post Recommend Mitigation",
        "description": "Generates allowlisted mitigation recommendations with evidence citations.",
        "operationId": "post_recommend_mitigation_api_v1_incidents__incident_id__recommend_mitigation_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MitigationPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/simulate-mitigation": {
      "post": {
        "summary": "Post Simulate Mitigation",
        "description": "Executes 3-way comparative verification (Baseline vs Unmitigated vs Mitigated).",
        "operationId": "post_simulate_mitigation_api_v1_incidents__incident_id__simulate_mitigation_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MitigationPlan"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Plan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MitigationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analyst/explain": {
      "post": {
        "summary": "Post Analyst Explain",
        "description": "Generates structured incident explanation using LLM or deterministic template fallback.",
        "operationId": "post_analyst_explain_api_v1_analyst_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalystContext"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalystExplanation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyst/ask": {
      "post": {
        "summary": "Post Analyst Ask",
        "description": "Answers operator questions with simulation-first verification of proposed actions.",
        "operationId": "post_analyst_ask_api_v1_analyst_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalystQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalystQuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalystContext": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "classification": {
            "$ref": "#/components/schemas/ClassificationClass"
          },
          "likely_cause": {
            "type": "string",
            "title": "Likely Cause"
          },
          "certainty": {
            "$ref": "#/components/schemas/CertaintyBand"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "affected_components": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Components"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "run_id",
          "classification",
          "likely_cause",
          "certainty",
          "risk_level",
          "evidence",
          "affected_components",
          "recommended_actions",
          "model_version"
        ],
        "title": "AnalystContext"
      },
...
      "AnalystQuestionRequest": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "question"
        ],
        "title": "AnalystQuestionRequest"
      },
...
      "CertaintyBand": {
        "type": "string",
        "enum": [
          "HIGH",
          "MEDIUM",
          "LOW"
        ],
        "title": "CertaintyBand"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImpactResult": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "projected_steps": {
            "type": "integer",
            "title": "Projected Steps"
          },
          "unmitigated_metrics": {
            "$ref": "#/components/schemas/VerificationMetrics"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance",
            "default": "CALCULATED"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "projected_steps",
          "unmitigated_metrics",
          "summary"
        ],
        "title": "ImpactResult"
      },
      "Incident": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id",
            "description": "e.g. GS-0001"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "scenario_ref": {
            "type": "string",
            "title": "Scenario Ref"
          },
          "created_at_step": {
            "type": "integer",
            "title": "Created At Step"
          },
          "created_at_wall": {
            "type": "string",
            "title": "Created At Wall"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "affected_components": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Components"
          },
          "classification": {
            "$ref": "#/components/schemas/ClassificationClass"
          },
          "attribution": {
            "$ref": "#/components/schemas/Attribution"
          },
          "certainty": {
            "$ref": "#/components/schemas/CertaintyBand"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskAssessment"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "recommended_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MitigationPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "mitigation_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MitigationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance",
            "default": "CALCULATED"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "run_id",
          "scenario_ref",
          "created_at_step",
          "created_at_wall",
          "status",
          "affected_components",
          "classification",
          "attribution",
          "certainty",
          "risk",
          "evidence",
          "model_version"
        ],
        "title": "Incident"
      },
...
      "MitigationPlan": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/MitigationInstruction"
            },
            "type": "array",
            "title": "Actions"
          },
          "evidence_citations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Citations"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance",
            "default": "CALCULATED"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "incident_id",
          "actions"
        ],
        "title": "MitigationPlan"
      },
      "MitigationResult": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "baseline_metrics": {
            "$ref": "#/components/schemas/VerificationMetrics"
          },
          "unmitigated_impact_metrics": {
            "$ref": "#/components/schemas/VerificationMetrics"
          },
          "mitigated_metrics": {
            "$ref": "#/components/schemas/VerificationMetrics"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance",
            "default": "CALCULATED"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "incident_id",
          "success",
          "summary",
          "baseline_metrics",
          "unmitigated_impact_metrics",
          "mitigated_metrics"
        ],
        "title": "MitigationResult"
      },
...
      "ScenarioSpec": {
        "properties": {
          "scenario_type": {
            "$ref": "#/components/schemas/ScenarioType",
            "default": "NORMAL"
          },
          "target_component": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Component"
          },
          "parameter_value": {
            "type": "number",
            "title": "Parameter Value",
            "default": 0.0
          },
          "start_step": {
            "type": "integer",
            "title": "Start Step",
            "default": 10
          },
          "duration_steps": {
            "type": "integer",
            "title": "Duration Steps",
            "default": 100
          },
          "total_steps": {
            "type": "integer",
            "title": "Total Steps",
            "default": 200
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          },
          "attack": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AttackSpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ScenarioSpec"
      },
      "ScenarioType": {
        "type": "string",
        "enum": [
          "NORMAL",
          "LOAD_INCREASE",
          "LINE_FAILURE",
          "GENERATOR_FAILURE",
          "VOLTAGE_INSTABILITY",
          "FREQUENCY_DISTURBANCE"
        ],
        "title": "ScenarioType"
      },
...
    }
  }
}